Moved functionality to seperate JS file
