+++ source/js/buyers-guide/template-js-handler/index.js
@@ -1,5 +1,6 @@
 import mobileNavStickinessHandler from "./mobile-nav-stickiness-handler";
 import mobileSearchBar from "./mobile-search-bar";
+import desktopSearchBar from "./desktop-search-bar";
 import diveDeeperListExpansionHandler from "./product-page-dive-deeper-list";
 import categoryDropdown from "./pni-category-dropdown";
 import productCommentGaEventHandler from "./product-page-comment-handler";
@@ -10,6 +11,7 @@ import productCommentGaEventHandler from "./product-page-comment-handler";
 export const bindEventHandlers = () => {
   mobileNavStickinessHandler();
   mobileSearchBar();
+  desktopSearchBar();
   diveDeeperListExpansionHandler();
   categoryDropdown();
   productCommentGaEventHandler();
